requirements.txt
fastapi
python-jose
sqlalchemy
passlib
bcrypt
python-multipart